Changed equals and hashCode in domain classes. Renamed folders in the persistence layer. The return value on all findById methods is wrapped in Optional
