+++ src/main/java/com/solvd/persistence/repositories/AddressRepository.java
@@ -3,10 +3,11 @@ package com.solvd.persistence.repositories;
 import com.solvd.domain.Address;
 
 import java.util.List;
+import java.util.Optional;
 
 public interface AddressRepository {
     void create(Address address);
     void deleteById(long id);
-    Address findById(long id);
+    Optional<Address> findById(long id);
     List<Address> getAll();
 }
