+++ src/main/java/com/solvd/persistence/repositories/ClientRepository.java
@@ -3,10 +3,11 @@ package com.solvd.persistence.repositories;
 import com.solvd.domain.Client;
 
 import java.util.List;
+import java.util.Optional;
 
 public interface ClientRepository {
     void create(Client client);
     void deleteById(long id);
-    Client findById(long id);
+    Optional<Client> findById(long id);
     List<Client> getAll();
 }
