+++ src/main/java/com/solvd/persistence/repositories/mybatisImpl/AddressRepositoryMybatisImpl.java
@@ -1,4 +1,4 @@
-package com.solvd.persistence.mybatisImpl;
+package com.solvd.persistence.repositories.mybatisImpl;
 
 import com.solvd.domain.Address;
 import com.solvd.persistence.Config;
@@ -6,6 +6,7 @@ import com.solvd.persistence.repositories.AddressRepository;
 import org.apache.ibatis.session.SqlSession;
 
 import java.util.List;
+import java.util.Optional;
 
 public class AddressRepositoryMybatisImpl implements AddressRepository {
     @Override
@@ -25,7 +26,7 @@ public class AddressRepositoryMybatisImpl implements AddressRepository {
     }
 
     @Override
-    public Address findById(long id) {
+    public Optional<Address> findById(long id) {
         try (SqlSession sqlSession = Config.getSessionFactory().openSession(true)) {
             AddressRepository addressRepository = sqlSession.getMapper(AddressRepository.class);
             return addressRepository.findById(id);
