+++ src/test/java/ClientTest.java
@@ -1,5 +1,5 @@
 import com.solvd.domain.Client;
-import com.solvd.persistence.mybatisImpl.ClientRepositoryMybatisImpl;
+import com.solvd.persistence.repositories.mybatisImpl.ClientRepositoryMybatisImpl;
 import com.solvd.persistence.repositories.ClientRepository;
 import com.solvd.service.ClientService;
 import com.solvd.service.impl.ClientServiceImpl;
